deduplicate message sending
